requirements.txt
lxml==4.1.1
click==6.7
numpy==1.14.0
requests==2.18.1
Pillow==5.0.0